requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
requests>=2.32.0,<3.0
APScheduler>=3.10.4,<4.0
python-dotenv>=1.0.1,<2.0
pypdf>=5.2.0,<6.0
python-multipart>=0.0.9,<1.0
